Added in loading of the memory regions from file into `mem_init`.
The simulator's behavior now matches the verilog simulator's. For a
given program, the simulator expects it to have one hex file per memory
region. The format of this file is an ASCII hex file, with one
hexadecimal word (32-bit value) per line.

The code now parses each of these files and loads it into memory.