Added in compilation of RISC-V test programs into the Makefile.
The assemble target will now read the variable `TEST` specified by the
user on the command line, and assemble the program. First, it
cross-compiles the program with the RISC-V runtime startup file. Then,
it copies out the user and kernel text and data sections. Last, it
converts these binary files to ASCII text files.

The Makefile was also generally cleaned up, and a runtime file was
added. This simply sets up the stack, invokes `main`, then uses `ECALL`
to terminate the program.