Updated the startup `crt0.S` file so that it doesn't use `li`.
The `li` instruction is a pseudoinstruction, so the assembler will
expand it out into actual instructions to accomplish its behavior. This
can be confusing to students first starting, because one line of
assembly code may generate several instructions. Thus, for simplicity,
it instead just uses `lui` to initialize the stack pointer.