Fixed an endianness issue when using `hexdump`.
The problem is that the format specifier for `hexdump`. It was
processing bytes one at a time (in chunks of 4), leading to each word
being laid out in little-endian order in the hex file. Instead, we want
to process the binary file 4 words at a time, so it is laid out in
big-endian order. This allows the simulator to parse the file easily.