Updated the crt0 startup file in several ways.
First, the stack address is now indicated as the ending address, as the
RISC-V ABI mandates that the stack grows downward. Also, the `call`
pseudoinstruction was used to invoke `main`, as this creates `jalr`,
meaning one less instruction needs to be implemented in the base starter
code. Last, the `ecall` that terminates the program uses the return
value from main as its argument (register a0 or x10).