Removed the dependency on a startup file for assembly tests.
The assembly tests now can run entirely on their own, without a startup
file. This was done so that the `jal` and `jalr` instructions would not
need to be implemented in the starter code, which would make the
pipeline lab tricky.

Now, the simulator handles initializing the stack pointer to the correct
value, which was the main purpose of the startup code. The C programs
naturally still use the startup file.