Fixed the parsing behavior for parsing 32-bit integers.
If the 32-bit value is being parsed, then it is treated like integer
constant, using the value 0 for `base` argument for `strtol`. Thus, for
a hexadecimal value to be used, it must be preceded by `0x` or `0X`.

Also, changed the formatting of hex files so the hexadecimal constant
produced have lowercase letters.