Fixed a bug with parsing 32-bit integer values.
The was how the code was treating the parsed `long` value. For parsing
32-bit integers, the desired behavior was to have signed decimal value,
and either signed or unsigned hexadecimal values. However, since the
code was not treating hex values specially, valid hex values like
`0xffffffff` were being rejected.

The code now handles this case, and also the parsing of values from the
hex files was changed to only permit unsigned values.