Switched the printing of memory to be big-endian.
Memory is now displayed in big-endian order, but the offsets from the
base address are explicitly listed to make it clear to the user. Also,
the `print_separator` function was used in other appropriate places, and
the printing of the CPU state in the register dump was moved to separate
function.