Fixed an issue with checking if a test exists.
The issue was that implicit rules were being used to match the test when
is had a *.s, *.S, or *.c extension. However, if the test didn't have
one of the extensions, then the check to see if the test would never
run. Thus, instead the cryptic 'no rule to make' message would be
displayed instead.