Updated the RISC-V ISA definitions to mirror the Verilog simulator's.
This was done to keep the two codebases consistent. The opcode and
function code enumerations had their typenames simplified, with the
'rsicv_' prefixed removed. Also, enumerations were moved into new types
to prevent duplicate values, which mirror the Verilog rule on this. A
minor comment update was done to the Verilog simulator as well.