Changed all references to 'MIPS' to 'RISC-V'.
This is to reflect the fact that we are now using a RISC-V core instead
of a MIPS core. In some cases, such as for the DC shell script, the
names needed to be updated to reflect new module names (e.g. the top
module). Also, the MIPS defines file was cleaned up so it only contains
syscall and exception unit relevant definitions, and its name was
changed as well.