Updated the R-type instructions to be better handled.
The previous way the enumerations were setup was confusing. In order to
properly decode an R-type instruction, you first need to look at the
7-bit function code to determine if it's a normal integer instruction,
and multiply instruction, or an "alternate" integer instruction (`sub`
and `sra` are the only two). Then, you use the 3-bit function code next
to determine which instruction it is.

Also, the starter code was updated to reflect the new format of the
enumerations.