Fixed a slight issue with compiling C test files.
The order in which the files are specified to the compiler affect how
they are laid out in the .text section. Before, the test was specified
first, then the startup file, so the entrypoint (`_start`), was not at
the beginning of the .text section, which the simulator requires. This
was fixed by switching the order of the files.