Removed the default values for the `TEST` and `TESTS` variables.
A fair bit of confusion to students was caused by the closeness of the
`TEST` versus `TESTS` variables. Students would mistakenly use `TESTS`
when `TEST` was expected, or vice-versa, leading to the wrong test
running, since the variable would default to a value.

By getting rid of the default, students will now be forced to always
specify the variable, so they will understand the different use cases
for each. The Makefile also has a check added to generate an error if
the user doesn't specify the appropriate variable.