Changed the build system so binary files are used instead of hex.
Back when the architecture was MIPS, the gcc compiler was not used to
assemble files, so there was no easy way to get disassembly files. Thus,
hex files were used to allow students to read what was assembled to aid
with debugging their code.

Now, since disassembly files can be generated, hex files are only an
unnecessary complication. It is much simpler to load the data directly
from binary files, so this extra step was removed from the code.