Fixed an issue with the verify command.
Since the `SHELL` variable was not specified, in defaulted to a minimal
`sh` shell, instead of `bash`. Thus, the idiom `&>` was not recognized
as redirecting stdout and stderr to file. Instead, the `&` was
interpreted as sending the command to execute in the background, leading
to the `sdiff` being printed after the Makefile finished executing.