Updated the startup file to capture the return value.
As RISC-V allows for double word return values, the return value from a
function can appear in both `a0` and `a1`, for the lower and upper
halves, respectively. Thus, to get these value in the register dump, two
registers are needed.

The startup file was updated to accommodate this. Also, we don't want to
clobber registers used by the `main` and other functions, so the return
values are stored in `sp` and `gp`, as these registers will be the same
values before and after calling `main`. Since `a0` is clobbered by the
`ecall` instruction needed to halt the simulator, we need to use another
register.