Changed the disassembly format and kept around the ELF file.
Since there are less files now that the hex files have gone away, we
should also keep around the compiled executable (ELF) file for assembled
programs.

Also, the disassembly format was changed so that no pseudoinstructions
are used in it, and it only uses the ISA names (not ABI names) for
registers. This makes it more useful for debugging as students will
typically see numeric register names, and will only see the actual
instructions assembled.