Added a compiler flag to restrict instructions to RV32IM.
A compiler flag has been added to the RISC-V compiler that restricts the
permitted instructions to the 32-bit integer and multiplication subsets
of the RISC-V ISA (RV32IM).

This will prevent students from using instructions that aren't
permitted, and from writing C programs that accidentally use
instructions outside this subset.