Removed `.s` files as allowed test files.
All test files must now be either `.c` or `.S` files. This is mainly to
simplify the test files students can create, and to avoid confusion with
the `.disassembly.s` files generated by the Makefile. The Makefile now
also has an explicit check to make sure the specified test has the
appropriate extension.