Changed the `verify` target and added an `autograde` one.
The new `autograde` target runs verification on a series of tests
specified by the user, or the set of tests students are required to pass
for the current lab. Unlike the old `verify` target, this command does
not show the output of individual tests, but rather prints a summary of
which tests passed and failed.

The `verify` target was changed to run on only a single test, since
`autograde` now handles this. The `verify` target shows the output of
the tests it is running. It generally isn't useful to show the output of
several tests, as users will typically debug one test at a time, so this
configuration makes sense.