General cleanup to the simulator code.
The major change was the style of the header comments for functions.
Instead of repeating the function name in every comment, which is very
much not necessary, the comments simply now only contain information
about the function. Also, title comments were added for includes, and
the author dates were updated.