Added an interface to abstract away the processor register file.
A register read and register write functions were added to abstract away
the details of interacting with the processor's register file. The main
reason for this was to emulate how memory is treated. Additionally, this
interface handles never allowing writes to register 0, as per the RISC-V
ISA specification.

Also, the code was generally cleaned up.