Removed the use of macros whenever possible.
This deals exclusively with constants in the program. Whenever possible,
and macro constant was replaced with a `static const <type>` definition
instead. Macros for constants are only required if they are used as
initializer elements (e.g. initializing another constant value). It is
generally better practice to have actual variables.