Cleaned up how memory segments are defined and handled.
The definitions of memory segments were moved to a separate file, and
the definitions were packed into a single array declaration. Also, the
CPU state now no longer copies in memory segments from the global
values, but instead assigns a pointer to the global array in its
structure. This simplifies a fair bit of the logic.