Added a verbose mode to the simulator.
When verbose mode is active, the simulator prints out a register dump
after every cycle of simulation that is run. Naturally, the user can
toggle this mode from the simulator shell with the new `verbose`
command. This can be useful to students for debugging, because it allows
for a way to perform a cycle-by-cycle diff.