Added in proper handling of global variables for C test programs.
When a C program is compiled, global variables can end up in a variety
of places. For optimizations, small global variables (e.g. ints and
other types), are placed in either the .sdata or .sbss section,
regardless of whether or not they're read-only and depending on if they
are initialized. Large global variables (e.g. strings and arrays) that
are read-only are placed in the .rodata section, while writable global
variables are placed into the .data section. Large initialized global
variables are placed in the .bss section.

To keep loading the program into memory for simulation simple, these
extra data sections are packed into either the .data or .text
section, depending on if they are read-only or not. This is accomplished
via a linker script, which also handles laying out the data at physical
locations in memory. The linker script in cleaner in general, since we
have specific requirements where the ELF sections go in memory, and this
is easier accomplished than with command-line options.