Fixed a bug with the `mdump` command, and changed its semantics.
The `mdump` command now prints out the memory address sequentially,
rather than treating each 4-byte sequence as a word, and printing it out
from MSB to LSB. Also, the ending address was changed to be exclusive,
as this tends to be more intuitive when typing in a range of memory
addresses.

When the range was inclusive `mdump` would crash if the starting and
ending addresses were equal due to an assertion failure. Now that the
ending address is exclusive, the check was updated so that the starting
address must be strictly less than the ending address.