Added proper handling of uninitialized global variables in C programs.
The issue was that the build system was not properly handling the .bss
section. When the program is compiled, any uninitialized and
zero-initialized global variables are placed in the .bss section. The
section only has a size, since all values are the same.

Thus, to convert this into binary format, extra flags must be added to
`objcopy` to tell it to expand the .bss section. The .bss section is
treated as its own section, and then it is concatenated onto the data
section's binary file when it is generated.