Changed how the syscall instruction is handled in the starter code.
Before, the (semi) complete implementation was given. Now, it was
changed so that the simulator will always halt if a syscall is seen,
which is not the correct behavior. The students must update the code to
get the correct behavior. This was done to match how the Verilog starter
code handles it.