Cleaned up how memory dumps are printed in the simulator.
When trying to find an address in the processor, the segment is now
returned, instead of a pointer to the segment's buffer. This allows for
the segment's name to be printed out in memory dumps. Also, memory dumps
now properly handle unaligned addresses, and addresses that are near the
boundary of a segment.