Replaced the Fibonacci test with a matrix multiplication one.
The Fibonacci test under the `447inputs` directory is redundant because
it is nearly identical to the `fibr` test under the `benchmarks`
directory. Thus, it was replaced with a matrix multiplication test. The
test shows how to write a C test, and also showcases how C tests can
utilize libgcc for emulating integer multiplication and floating point
instructions.