Changed how 7-bit function codes are handled for the ISA.
In the RISC-V ISA, there are several 3-bit function code which map to
more than one instruction (add/sub, srl/sra, and srli/srai). These
instructions are distinguished by the 7-bit function codes. These
function codes are shared by both I-type and R-type instructions.

The code before treated the 7-bit function code as separate for the
I-type and R-type instructions, which it no longer does. The code was
updated appropriately for this. Also, the names of the enumerations were
made consistent.