Suppressed a warning when `ldconfig` cannot be found.
The Makefile uses `ldconfig` to check if the readline library is on the
system. This is done for the sake of convenience, so if `ldconfig` is
not present on the system, then the Makefile should skip this check.
This was happening when the Makefile was run in non-interactive mode, as
the `PATH` variable was not yet setup to include the `ldconfig` utility.