Changed the fibm test slightly.
The local array used by the Fibonacci function now uses a static local
variable. This is to ensure that the memory writes do not get optimized
out. Also, this is done so that the array for the Fibonacci sequence can
be easily viewed by the students as it is now located in the data
segment (.bss section).